# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: UniClub API
  version: 1.0.0
  description: |
    유니클럽 독서모임 플랫폼 API 문서입니다.
    인증이 필요한 API는 NextAuth 세션 쿠키를 통해 인증됩니다.
servers:
  - url: https://club.bestcome.org/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
...
paths:
...
  /club/timeline:
    get:
      summary: 활동 타임라인
      tags: [Club]
      security:
        - cookieAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, report, attendance, badge]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineItem'
                  nextCursor:
                    type: string
                    nullable: true
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth 세션 쿠키
  schemas:
...
    TimelineItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [report, attendance, badge, quote]
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time